requirements.txt
streamlit==1.47.1
langchain==0.3.27
faiss-cpu==1.11.0
python-dotenv==1.1.1
langchain-community==0.3.27
tiktoken==0.9.0
sentence-transformers==5.0.0